+++ src/view/GridViewConsole.cpp
@@ -20,20 +20,20 @@ void GridViewConsole::Display()
     if(is_wider_than_10)
     {
         output << setw(row_indicator_max_w + 1) << "";
-		for(int i = 0; i < grid.W; i+=10) output << setw(10) << left << i;
+		for(size_t i = 0; i < grid.W; i+=10) output << setw(10) << left << i;
 		output << endl;
     }
 
 	output << setw(row_indicator_max_w + 1) << "";
-	for(int i = 0; i < grid.W; i++) output << i % 10;
+	for(size_t i = 0; i < grid.W; i++) output << i % 10;
 	output << endl;
 
 	output << setw(row_indicator_max_w) << "" << "+" << setfill('-') << setw(grid.W) << "" << "+" << setfill(' ') << endl;
 
-    for(int i = 0; i < grid.H; i++)
+    for(size_t i = 0; i < grid.H; i++)
 	{
 		output << setw(row_indicator_max_w) << right << i << "|";
-		for(int j = 0; j < grid.W; j++)
+		for(size_t j = 0; j < grid.W; j++)
 		{
 			int f = i * grid.W + j;
             if(grid.is_flag[f]) output << "x";
@@ -48,13 +48,13 @@ void GridViewConsole::Display()
 	output << setw(row_indicator_max_w) << "" << "+" << setfill('-') << setw(grid.W) << "" << "+" << setfill(' ') << endl;
 
 	output << setw(row_indicator_max_w + 1) << "";
-	for(int i = 0; i < grid.W; i++) output << i % 10;
+	for(size_t i = 0; i < grid.W; i++) output << i % 10;
 	output << endl;
 
     if(is_wider_than_10)
     {
         output << setw(row_indicator_max_w + 1) << "";
-		for(int i = 0; i < grid.W; i+=10) output << setw(10) << left << i;
+		for(size_t i = 0; i < grid.W; i+=10) output << setw(10) << left << i;
 		output << endl;
     }
 	//std::system("cls");  // this is very ugly
